Split 'data' out of web.py
