+++ saci/webui/web.py
@@ -1,6 +1,3 @@
-from dataclasses import dataclass
-from collections import defaultdict
-from enum import StrEnum
 import asyncio
 import logging
 import uuid
@@ -8,7 +5,6 @@ import uuid
 import importlib
 from pathlib import Path
 import os
-from typing import TypeVar
 
 import httpx
 
@@ -18,21 +14,13 @@ from fastapi.responses import FileResponse
 from websockets.asyncio.client import connect as ws_connect, ClientConnection as WsClientConnection
 from websockets.exceptions import InvalidStatus as WsInvalidStatus, ConnectionClosedOK as WsConnectionClosedOK
 
-from saci.modeling.annotation import Annotation
-from saci.modeling.device import ComponentID, ComponentBase
-from saci.modeling.device.control.controller import Controller
-from saci.modeling.device.sensor.compass import CompassSensor
-from saci.modeling.device.esc import ESC
-from saci.modeling.device.motor.steering import Steering
-from saci.modeling.device.webserver import WebServer
+import saci.webui.data as data
 from saci.modeling.state.global_state import GlobalState
-from saci.modeling.vulnerability.base_vuln import MakeEntryEffect, VulnerabilityEffect
-from saci.webui.web_models import AnalysisUserInfo, AnnotationID, AnnotationModel, BlueprintID, CPVResultModel, ComponentTypeID, ComponentTypeModel, DeviceModel, HypothesisID, HypothesisModel, ParameterTypeModel
+from saci.webui.web_models import AnalysisID, AnalysisUserInfo, AnnotationID, AnnotationModel, BlueprintID, CPVResultModel, ComponentTypeID, ComponentTypeModel, DeviceModel, ParameterTypeModel
+from saci_db.cpvs import CPVS
 
 from ..orchestrator import identify
 from ..deserializer import ingest
-from ..modeling import Device
-from ..modeling.device import Wifi, Motor, GPSReceiver
 
 l = logging.getLogger(__name__)
 
@@ -55,7 +43,7 @@ async def serve_frontend_root():
 def get_blueprints() -> dict[BlueprintID, DeviceModel]:
     # TODO: eventually we won't want to send all this data at once
     # TODO: store the hypotheses per-blueprint
-    return {bp_id: DeviceModel.from_device(bp, hypotheses[bp_id], annotations[bp_id]) for bp_id, bp in blueprints.items()}
+    return {bp_id: DeviceModel.from_device(bp, data.hypotheses[bp_id], data.annotations[bp_id]) for bp_id, bp in data.blueprints.items()}
 
 
 @app.post("/api/blueprints/{bp_id}")
@@ -65,9 +53,9 @@ def create_or_update_blueprint(bp_id: str, serialized: dict, response: Response)
         raise HTTPException(status_code=400, detail=err)
 
     try:
-        ingest(serialized, INGESTION_DIR / bp_id, force=True)
+        ingest(serialized, data.INGESTION_DIR / bp_id, force=True)
 
-        if bp_id in blueprints:
+        if bp_id in data.blueprints:
             # TODO: actually check to make sure this is actually a fast-forwarded version
             created = False
         else:
@@ -78,8 +66,8 @@ def create_or_update_blueprint(bp_id: str, serialized: dict, response: Response)
 
     importlib.invalidate_caches()
     # TODO: probably don't need to reload *all* the ingested modules
-    importlib.reload(ingested)
-    blueprints[bp_id] = ingested.devices[bp_id]
+    importlib.reload(data.ingested)
+    data.blueprints[bp_id] = data.ingested.devices[bp_id]
 
     if created:
         response.status_code = status.HTTP_201_CREATED
@@ -90,21 +78,21 @@ def create_or_update_blueprint(bp_id: str, serialized: dict, response: Response)
 @app.post("/api/blueprints/{bp_id}/annotation")
 def create_annotation(bp_id: str, annot_model: AnnotationModel) -> AnnotationID:
     # this AnnotationID selection is a stupid mock for until we get the database code in here :)
-    if bp_id not in blueprints:
+    if bp_id not in data.blueprints:
         raise HTTPException(status_code=404, detail="Blueprint not found")
 
     annot_id = str(uuid.uuid4())
-    while annot_id in annotations[bp_id]:
+    while annot_id in data.annotations[bp_id]:
         annot_id = str(uuid.uuid4())
 
-    annotations[bp_id][annot_id] = annot_model.to_annotation()
+    data.annotations[bp_id][annot_id] = annot_model.to_annotation()
 
     return annot_id
 
 
 @app.get("/api/blueprints/{bp_id}/cpvs")
 def identify_cpvs(bp_id: str) -> list[CPVResultModel]:
-    if (blueprint := blueprints.get(bp_id)) is None:
+    if (blueprint := data.blueprints.get(bp_id)) is None:
         raise HTTPException(status_code=404, detail="Blueprint not found")
 
     # TODO: re-introduce the queueing model once this takes more time
@@ -116,29 +104,14 @@ def identify_cpvs(bp_id: str) -> list[CPVResultModel]:
         for path in paths
     ]
 
-T = TypeVar('T')
-def _all_subclasses(c: type[T]) -> list[type[T]]:
-    return [c] + [subsubc for subc in c.__subclasses__() for subsubc in _all_subclasses(subc)]
-
-
-def component_type_id(comp_type: type[ComponentBase]) -> str:
-    return f"{comp_type.__module__}.{comp_type.__qualname__}"
-
-# TODO: so janky
-all_component_types: dict[ComponentTypeID, type[ComponentBase]] = {}
-for comp_type in _all_subclasses(ComponentBase):
-    type_id = component_type_id(comp_type)
-    if dup := all_component_types.get(type_id):
-        l.warning(f"duplicately-IDed component types {dup} and {comp_type} (both have ID {type_id}), only including {dup}")
-    all_component_types[type_id] = comp_type
-
 @app.get("/api/components/")
 def list_component_types() -> list[ComponentTypeID]:
-    return list(all_component_types)
+    return list(data.all_component_types)
+
 
 @app.get("/api/components/{type_id}")
 def component_type_details(type_id: str) -> ComponentTypeModel:
-    if (comp_type := all_component_types.get(type_id)) is None:
+    if (comp_type := data.all_component_types.get(type_id)) is None:
         raise HTTPException(status_code=404, detail="No component type with that ID")
     return ComponentTypeModel(
         # TODO: have component types have better human-readable names
@@ -150,11 +123,11 @@ def component_type_details(type_id: str) -> ComponentTypeModel:
 
 @app.post("/api/ingest_blueprint")
 def ingest_blueprint_legacy(name: str, serialized: dict, force: bool = False):
-    if name in blueprints:
+    if name in data.blueprints:
         return {"error": "exists"}, 400
     try:
         # TODO: move this to another thread and return a promise?
-        ingest(serialized, INGESTION_DIR / name, force=force)
+        ingest(serialized, data.INGESTION_DIR / name, force=force)
     except FileExistsError as e:
         return {"error": "exists"}, 400
     except ValueError as e:
@@ -163,11 +136,11 @@ def ingest_blueprint_legacy(name: str, serialized: dict, force: bool = False):
 
     importlib.invalidate_caches()
     # TODO: probably don't need to reload *all* the ingested modules
-    importlib.reload(ingested)
+    importlib.reload(data.ingested)
     # TODO: should probably just separate builtin and ingested blueprints...
     blueprint_id = name
-    device = ingested.devices[blueprint_id]
-    blueprints[blueprint_id] = device
+    device = data.ingested.devices[blueprint_id]
+    data.blueprints[blueprint_id] = device
 
     return {"id": blueprint_id, "name": device.name}
 
@@ -190,38 +163,18 @@ except httpx.ConnectError:
     l.warning("can't connect to app-controller, is it up and the URL configured correctly?")
 
 
-class InteractionModel(StrEnum):
-    UNKNOWN = "Unknown"
-    X11 = "X11"
-
-@dataclass(frozen=True)
-class Analysis:
-    """All the information associated with an analysis type, including what the system needs to know to launch it."""
-    user_info: AnalysisUserInfo
-    interaction_model: InteractionModel
-    images: list[str]
-
-    def as_appconfig(self):
-        return {
-            "name": "app",
-            "interaction_model": self.interaction_model,
-            "images": self.images,
-            "always_pull_images": False,
-        }
-
-AnalysisID = str
 
 @app.get("/api/blueprints/{bp_id}/analyses")
 def get_analyses(bp_id: str) -> dict[AnalysisID, AnalysisUserInfo]:
     # for now ignore bp_id, but eventually analyses will be available per-device or something.
     # return mapping of analysis ID to analysis info
-    return {id_: analysis.user_info for id_, analysis in analyses.items()}
+    return {id_: analysis.user_info for id_, analysis in data.analyses.items()}
 
 @app.post("/api/blueprints/{bp_id}/analyses/{analysis_id}/launch")
 async def launch_analysis(bp_id: str, analysis_id: str) -> int:
-    if analysis_id not in analyses:
+    if analysis_id not in data.analyses:
         raise HTTPException(status_code=400, detail="analysis not found")
-    analysis = analyses[analysis_id]
+    analysis = data.analyses[analysis_id]
     async with httpx.AsyncClient() as client:
         create_resp = await client.post(f"{APP_CONTROLLER_URL}/api/app", json=analysis.as_appconfig())
         if not create_resp.is_success:
@@ -279,131 +232,3 @@ async def vnc_proxy(*, websocket: WebSocket, app_id: int):
         if rest is not None:
             raise rest
 
-# delayed import
-from saci_db.devices import devices, ingested
-from saci_db.cpvs import CPVS
-
-if (dirname := os.getenv("INGESTION_DIR")) is not None:
-    INGESTION_DIR = Path(dirname)
-else:
-    INGESTION_DIR = Path(ingested.__file__).resolve().parent
-del dirname
-
-blueprints: dict[BlueprintID, Device] = devices | ingested.devices
-
-# TODO: this is hacky and an indication that we should have a better way of doing this...
-def _find_comps(device: Device, comp_type: type[ComponentBase]) -> list[ComponentID]:
-    return [comp_id for comp_id, comp in device.components.items() if isinstance(comp, comp_type)]
-
-def _find_comp(device: Device, comp_type: type[ComponentBase]) -> ComponentID:
-    comps = _find_comps(device, comp_type)
-    if len(comps) == 0:
-        raise ValueError(f"device {device!r} has no component of type {comp_type}")
-    elif len(comps) > 1:
-        raise ValueError(f"device {device!r} has more than one component of type {comp_type}")
-    else:
-        return comps[0]
-
-rover = blueprints["ngcrover"]
-
-analyses: dict[AnalysisID, Analysis] = {
-    "taveren_model": Analysis(
-        user_info=AnalysisUserInfo(
-            name="Model: Ta'veren Controller",
-            # TODO: hackyyyyy... should either use different controllers' different IDs (now that they have them!) or have some nice query mechanism
-            components_included=[
-                _find_comps(rover, WebServer)[0],
-                _find_comps(rover, Controller)[0],
-            ],
-        ),
-        interaction_model=InteractionModel.X11,
-        images=["taveren:latest"],
-    ),
-    "binsync_re": Analysis(
-        user_info=AnalysisUserInfo(
-            name="Model: BinSync-enabled RE",
-            components_included=[_find_comps(rover, Controller)[0]],
-        ),
-        interaction_model=InteractionModel.X11,
-        images=["ghcr.io/twizmwazin/app-controller/firefox-demo:latest"],
-    ),
-    "hybrid_automata": Analysis(
-        user_info=AnalysisUserInfo(
-            name="Model: Hybrid Automata",
-            components_included=_find_comps(rover, Controller) + [
-                _find_comp(rover, GPSReceiver),
-                _find_comp(rover, CompassSensor),
-                _find_comp(rover, Steering),
-                _find_comp(rover, ESC),
-                _find_comp(rover, Motor),
-            ],
-        ),
-        interaction_model=InteractionModel.X11,
-        images=["ghcr.io/twizmwazin/app-controller/firefox-demo:latest"],
-    ),
-    "gazebo_hybrid_automata": Analysis(
-        user_info=AnalysisUserInfo(
-            name="Co-Simulation: Gazebo + Hybrid Automata",
-            components_included=_find_comps(rover, Controller) + [
-                _find_comp(rover, GPSReceiver),
-                _find_comp(rover, CompassSensor),
-                _find_comp(rover, Steering),
-                _find_comp(rover, ESC),
-                _find_comp(rover, Motor),
-            ],
-        ),
-        interaction_model=InteractionModel.X11,
-        images=["ghcr.io/cpslab-asu/gzcm/px4/firmware:0.2.0", "ghcr.io/cpslab-asu/gzcm/px4/gazebo:harmonic"],
-    ),
-    "gazebo_firmware": Analysis(
-        user_info=AnalysisUserInfo(
-            name="Co-Simulation: Gazebo + Firmware",
-            components_included=_find_comps(rover, Controller) + [
-                _find_comp(rover, GPSReceiver),
-                _find_comp(rover, CompassSensor),
-                _find_comp(rover, Steering),
-                _find_comp(rover, ESC),
-                _find_comp(rover, Motor),
-            ],
-        ),
-        interaction_model=InteractionModel.X11,
-        images=["onex:latest"],
-    ),
-}
-
-hypotheses: dict[BlueprintID, dict[HypothesisID, HypothesisModel]] = defaultdict(dict, {
-    "ngcrover": {
-        "webserver_stop": HypothesisModel(
-            name="From the webserver, stop the rover.",
-            entry_component=_find_comp(rover, Wifi),
-            exit_component=_find_comp(rover, Motor),
-        ),
-        "emi_compass": HypothesisModel(
-            name="Using EMI, influence the compass to affect the mission.",
-            entry_component=_find_comp(rover, CompassSensor),
-            exit_component=None,
-        ),
-        "wifi_rollover": HypothesisModel(
-            name="Over WiFi, subvert the control system to roll the rover.",
-            entry_component=_find_comp(rover, Wifi),
-            exit_component=_find_comp(rover, Steering),
-        ),
-    },
-})
-
-annotations: dict[BlueprintID, dict[AnnotationID, Annotation]] = defaultdict(dict, {
-    "ngcrover": {
-        "wifi_open": Annotation(
-            attack_surface=ComponentID("wifi"),
-            effect=MakeEntryEffect(reason="wifi is open", nodes=frozenset([ComponentID("wifi")])),
-            underlying_vulnerability=None,
-            attack_model="connect to the AP without creds",
-        ),
-        "hidden_stop": Annotation(
-            attack_surface=ComponentID("webserver"),
-            effect=VulnerabilityEffect(reason="hidden stop command in the webserver"),
-            underlying_vulnerability=None,
-            attack_model="hit the stop endpoint on the webserver",
-        ),
-    },
-})
